+++ go/kryon/style_runtime_test.go
@@ -77,7 +77,7 @@ func TestButtonWithoutStylePackHasNoVisualDefaults(t *testing.T) {
 			style.Focus != 0 || style.Radius != 0 || style.BorderWidth != 0 {
 			t.Fatalf("unstyled button leaked visual defaults: %+v", style)
 		}
-		if style.Fields&uint32(StyleMaterial) == 0 || style.Material != int32(MaterialFlat) {
+		if style.Fields&uint32(StyleMaterial) == 0 || style.Material != MaterialFlat {
 			t.Fatalf("unstyled button must use flat material fallback: %+v", style)
 		}
 		if style.Opacity != 1 || style.FontSize == 0 {
